# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "KnowChain LLM API",
      "version": "1.0.0"
    },
    "paths": {
      "/chat/web": {
        "post": {
          "summary": "Chat using indexed website documents",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessionId": { "type": "string" },
                    "query": { "type": "string" }
                  },
                  "required": ["sessionId", "query"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Chat response",
              "content": { "application/json": {} }
            }
          }
        }
      },
      "/chat/pdf": {
        "post": {
          "summary": "Chat using indexed PDF documents",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessionId": { "type": "string" },
                    "query": { "type": "string" },
                    "collectionName": { "type": "string" }
                  },
                  "required": ["sessionId", "query"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Chat response",
              "content": { "application/json": {} }
            }
          }
        }
      },
      "/chat/text": {
        "post": {
          "summary": "Chat using indexed text documents",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessionId": { "type": "string" },
                    "query": { "type": "string" },
                    "collectionName": { "type": "string" }
                  },
                  "required": ["sessionId", "query"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Chat response",
              "content": { "application/json": {} }
            }
          }
        }
      },
      "/chat/index/web": {
        "post": {
          "summary": "Index a website for chat",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": { "type": "string" }
                  },
                  "required": ["url"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Indexing response",
              "content": { "application/json": {} }
            }
          }
        }
      },
      "/chat/index/pdf": {
        "post": {
          "summary": "Index a PDF file for chat",
          "requestBody": {
            "required": true,
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "file": { "type": "string", "format": "binary" },
                    "collectionName": { "type": "string" }
                  },
                  "required": ["file"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Indexing response",
              "content": { "application/json": {} }
            }
          }
        }
      },
      "/chat/index/text": {
        "post": {
          "summary": "Index raw text for chat",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "text": { "type": "string" },
                    "collectionName": { "type": "string" }
                  },
                  "required": ["text"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Indexing response",
              "content": { "application/json": {} }
            }
          }
        }
      }
    }
  }
  